Refactor: remove deprecated files and update imports for CustomSafeLoader
